Move AdminHomeController to its proper home
in the web.controller.admin package.
