+++ srcalc/src/test/java/gov/va/med/srcalc/web/controller/admin/AdminHomeControllerIT.java
@@ -1,4 +1,4 @@
-package gov.va.med.srcalc.web.controller;
+package gov.va.med.srcalc.web.controller.admin;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
