+++ srcalc/src/test/java/gov/va/med/srcalc/web/controller/admin/EditVariableControllerIT.java
@@ -10,7 +10,6 @@ import gov.va.med.srcalc.domain.model.*;
 import gov.va.med.srcalc.service.AdminService;
 import gov.va.med.srcalc.test.util.IntegrationTest;
 import gov.va.med.srcalc.test.util.TestHelpers;
-import gov.va.med.srcalc.web.controller.AdminHomeController;
 import gov.va.med.srcalc.web.view.Views;
 import gov.va.med.srcalc.web.view.admin.*;
